+++ crates/gpui/src/platform/keyboard.rs
@@ -2,6 +2,14 @@ use std::borrow::Cow;
 
 use super::{Keystroke, WindowsKeyboardMapper};
 
+/// A trait for platform-specific keyboard layouts
+pub trait PlatformKeyboardLayout {
+    /// Get the keyboard layout ID, which should be unique to the layout
+    fn id(&self) -> &str;
+    /// Get the keyboard layout display name
+    fn name(&self) -> &str;
+}
+
 /// TODO:
 pub trait KeyboardMapper {
     /// TODO:
