+++ crates/zed/src/zed.rs
@@ -1225,7 +1225,7 @@ pub fn handle_keymap_file_changes(
     .detach();
 
     // let mut current_mapping = settings::get_key_equivalents(cx.keyboard_layout().id());
-    let mut keyboard_layout = cx.keyboard_layout().to_string();
+    let mut keyboard_layout = cx.keyboard_layout().id().to_string();
     cx.on_keyboard_layout_change(move |cx| {
         let current_layout = cx.keyboard_layout().id().to_string();
         if current_layout != keyboard_layout {
